# openapi-generation-demo/src/main/resources/openapi/openapi.json
{"openapi":"3.1.0","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/v2/demo-api/validation/{value}":{"post":{"tags":["demo-controller"],"operationId":"validatePathVariable","parameters":[{"name":"value","in":"path","required":true,"schema":{"type":"integer","format":"int32","minimum":0}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DemoResponse"}}}}}}},"/v2/demo-api/new-complex":{"post":{"tags":["demo-controller"],"operationId":"createComplexObject","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DemoComplexObject"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DemoResponse"}}}}}}}},"components":{"schemas":{"DemoResponse":{},"DemoComplexObject":{"properties":{"bigDecimal":{"type":"number","description":"A BigDecimal","example":1111111.1234,"x-field-extra-annotation":"@jakarta.validation.constraints.PositiveOrZero"},"bigInteger":{"type":"integer","format":"biginteger","description":"A BigInteger"},"string":{"type":"string","description":"A String","maxLength":255,"minLength":1,"pattern":"[a-z]+"},"longInteger":{"type":"integer","format":"int64","description":"A long","maximum":100,"minimum":0},"integer":{"type":"integer","format":"int32","description":"An int"},"aShort":{"type":"integer","format":"int32","description":"A short"},"aByte":{"type":"string","format":"byte","description":"A byte"},"uuid":{"type":"string","format":"uuid","description":"A uuid"},"instant":{"type":"string","format":"date-time","description":"An Instant","example":"2022-02-15T18:35:24.00Z"},"list":{"type":"array","description":"A List<InnerObject>","items":{"$ref":"#/components/schemas/InnerObject"}},"set":{"type":"array","description":"A Set<InnerObject>","items":{"$ref":"#/components/schemas/InnerObject"},"uniqueItems":true},"map":{"type":"object","additionalProperties":{"$ref":"#/components/schemas/InnerObject"},"description":"A Map<String, InnerObject>"}},"required":["bigDecimal","bigInteger","instant","list","map","set","string"]},"InnerObject":{"properties":{"integer":{"type":"integer","format":"int32"},"string":{"type":"string"}}}}}}